requirements.txt
passlib==1.7.2
SQLAlchemy==1.3.13
starlette==0.12.9
uvicorn==0.11.3
pydantic==1.4
requests==2.21.0
namedlist==1.7
PyExecJS==1.5.1
fastapi==0.52.0
PyJWT==1.7.1
bcrypt==3.1.7